Download old docker versions

+++ init.sh
@@ -48,8 +48,20 @@ export DOCKER_BINARY=${DATADIR}/docker-${DOCKER_VERSION}
 
 if [ ! -e ${DOCKER_BINARY} ] ; then
   curl -fsSL -o ${SKYPORT_TMPDIR}/docker-${DOCKER_VERSION}.tgz https://download.docker.com/linux/static/stable/x86_64/docker-${DOCKER_VERSION}.tgz 
+  if [ ! -e ${SKYPORT_TMPDIR}/docker-${DOCKER_VERSION}.tgz ] ; then
+    echo "download.docker.com did not work, try old location get.docker.com..."
+    curl -fsSL -o ${SKYPORT_TMPDIR}/docker-${DOCKER_VERSION}.tgz  https://get.docker.com/builds/Linux/x86_64/docker-${DOCKER_VERSION}.tgz
+  fi
+  
+  if [ ! -e ${SKYPORT_TMPDIR}/docker-${DOCKER_VERSION}.tgz ] ; then
+    echo "docker binary not found"
+    exit 1
+  fi
+  
   tar -xvzf ${SKYPORT_TMPDIR}/docker-${DOCKER_VERSION}.tgz -C ${SKYPORT_TMPDIR} docker/docker 
   mv ${SKYPORT_TMPDIR}/docker/docker ${DATADIR}/docker-${DOCKER_VERSION}
+  
+  
 fi
